+++ tests/nc.R
@@ -10,8 +10,10 @@ foo <- st %>% st_set_dimensions(1, st_geometry(nc)) # %>% st_set_dimensions(3, c
 foo
 frac = function(x) x[2] / x[1]
 frac2 = function(x) c(sidsr = x[2] / x[1], nwbr = x[3] / x[1])
+frac2an = function(x) c(x[2] / x[1], x[3] / x[1])
 st_apply(foo, c(1,3), frac)
 st_apply(foo, c(1,3), frac2)
+st_apply(foo, c(1,3), frac2an)
 library(abind)
 aperm(st_apply(foo, c(1,3), frac2), c(2,3,1))